Close the C_SclString migration residue, and let the compiler check printf formats (#30)
* Parse "0x" attributes as hex; repair parameter set file loading

The C_SclString migration's indexing residue is fixed; this is its parsing
residue. C_SclString::ToInt() was hex-aware. std::stoi is hard-wired to base 10:
std::stoi("0x10") reads the leading '0', stops at the 'x', and returns 0 without
throwing, so every try/catch around it stays quiet.

C_SclStringUtil.hpp already documents this exact failure and provides
ScanBaseCompat, which selects base 16 for a "0x" prefix without treating a
leading "0" as octal. These call sites were simply never moved onto it.

Seven of the eight XML attribute getters document "Can handle 0x notation to
interpret hex values", and each explains that it avoids XMLElement::Query
precisely because Query cannot parse hex -- and then calls std::stoi with no
base. Fixing three of them fixes all eight, because the four *Error variants
delegate to the plain getters.

What that broke is not hypothetical. Two production writers emit "0x" values and
both are on the parameter set format:

- C_OscChecksummedXml wrote the file CRC as "0x" + hex and read it back with
  GetAttributeUint32, so the CRC read as 0 and never matched. The class could not
  load a file it had just written.
- C_OscParamSetFilerBase wrote "0x0001" as the file version and read it with a
  base-10 std::stoi, got 0, and rejected the file as an unsupported version.

C_OscChecksummedXml is used only by the parameter set filers, so loading any
.syde_psi failed twice over. Every other filer writes its version with
std::to_string, so their readers were already correct.

Verified by round trip rather than by reading: save a checksummed file, load it
back, observe C_CHECKSUM. Reverting both fixes against the new tests fails 5 on a
clean build. The two negative controls -- DetectsTamperedContent and
UnsupportedVersionIsRejected -- pass in both states, so these fixes restore the
checks rather than disabling them. 308 tests pass.


* Fix 1-based loop bounds in the comment sanitiser and TSP name sanitiser

The third form of the same C_SclString residue. An inclusive bound against a
count -- "index <= length" -- is 1-based thinking applied to 0-based indexing.

C_OscUtils::h_NiceifyStringForCeComment never examined character 0, and its last
iteration read and then wrote c_Result[size()], which is undefined for the
non-const operator[].

That function exists to stop a Datapool comment from breaking the generated C
file it is embedded in, and position 0 is exactly where it could not look.
Reverting the fix against the new tests shows "*/rest", "\nbc", "`ackquote",
"@t" and "text\" all passing through untouched -- a comment terminator and a raw
newline going straight into a block comment.

The sibling immediately above it, h_NiceifyStringForFileName, is correct 0-based.
Same file, adjacent functions, one converted and one not.

opensyde_tsp_convert's mh_Sanitize had the same bounds, dropping the first
character and appending a '_' for the terminator: "ESX-4CS-GW" came out as
"SX_4CS_GW_" where its own doc comment says "ESX_4CS_GW".

test_osc_utils_niceify.cpp is new and covers every position, including the
lone-star and mid-string backslash cases that must be left alone. 312 tests pass.


* Let the compiler type-check printf-style format strings

PrintFormattedCompat is a varargs function with no format attribute, so neither
clang nor gcc checked a single one of its ~200 call sites. A wrong conversion or
a missing argument is undefined behaviour that no test reliably catches, and a
grep cannot see a type mismatch at all.

Adding __attribute__((format(printf, 1, 2))) hands the whole class to the
compiler, permanently and on every platform, instead of to a scan that has to be
remembered and re-run.

It immediately found four sites passing a 64-bit size_type to %02d, in the two
flash-progress messages. All four now cast to int32_t, matching the cast the same
file already uses a few lines further down.

A separate scan for argument-count mismatches across all 200 call sites found
none, once it was taught that C++ concatenates adjacent string literals -- the
first version of it reported eleven false positives for exactly that reason, and
was only trusted after it was checked against planted cases.


* Use PRIu64 for the CAN Monitor timestamp instead of a hardcoded "llu"

The format attribute caught this on the Linux build host and not on macOS, which
is the whole point of it: uint64_t is unsigned long on Linux and unsigned long
long on macOS and Windows, so no fixed length modifier is correct everywhere.
"%013llu" is right on two of the three platforms this project ships.

PRIu64 expands to whatever is correct for the target.

tinyxml2 is left alone -- it is vendored, and already casts to long long
explicitly for this same reason. C_Uti's "%lu" pairs with a Windows DWORD, which
is correct; the attribute will now confirm that on the Windows job rather than
leaving it unchecked.


* Record the format-attribute result in FINDINGS.md


---------

+++ libraries/opensyde_core/protocol_drivers/system_update/C_OscSuSequences.cpp
@@ -637,7 +637,7 @@ std::error_code C_OscSuSequences::m_FlashOneFileOpenSydeHex(const stw::hex_file:
          {
             std::string c_Text;
             c_Text = PrintFormattedCompat("Writing data for area %02d/%02d  byte %08u/%08u ...",
-                                  s32_Area + 1, orc_HexDataDump.at_Blocks.size(),
+                                  s32_Area + 1, static_cast<int32_t>(orc_HexDataDump.at_Blocks.size()),
                                   u32_AreaSize - u32_RemainingBytes, u32_AreaSize);
             q_Abort = m_ReportProgress(eUPDATE_SYSTEM_OSY_NODE_FLASH_HEX_AREA_TRANSFER_START, C_NO_ERR,
                                        u8_ProgressPercentage, mc_CurrentNode, c_Text);
@@ -707,7 +707,8 @@ std::error_code C_OscSuSequences::m_FlashOneFileOpenSydeHex(const stw::hex_file:
          //report "final" status:
          std::string c_Text;
          c_Text = PrintFormattedCompat("Writing data for area %02d/%02d  byte %08u/%08u ...",
-                               s32_Area + 1, orc_HexDataDump.at_Blocks.size(), u32_AreaSize, u32_AreaSize);
+                               s32_Area + 1, static_cast<int32_t>(orc_HexDataDump.at_Blocks.size()), u32_AreaSize,
+                               u32_AreaSize);
          (void)m_ReportProgress(eUPDATE_SYSTEM_OSY_NODE_FLASH_HEX_AREA_TRANSFER_START, C_NO_ERR,
                                 u8_ProgressPercentage, mc_CurrentNode, c_Text);
